\name{generate_cluster_cor_mat}
\alias{generate_cluster_cor_mat}
\title{ Generate Cluster Correlation Matrix}
\description{ 
  \code{generate_cluster_cor_mat} generates the cluster correlation matrix 
  to examine the within- and between-cluster correlations.  
}

\usage{
 generate_cluster_cor_mat(observed_cor_matrix, cluster_vector) 
}

\arguments{
  \item{observed_cor_matrix}{ observed correlation matrix }
  \item{cluster_vector}{ vector of clusters }
}
\value{
  \item{cluster_cor_mat}{a cluster correlation matrix }}

\author{Mike Nowak \email{michael.nowak@gmail.com}}


\examples{
##


}
\keyword{ ~sna }

