\name{permute_matrix}
\alias{permute_matrix}
\title{ Triad census for networks }
\description{ 
  \code{permute_matrix} permute the network to examine the within- and between-cluster correlations.  
}

\usage{
permute_matrix(mem_vector, adj_matrix)
}

\arguments{
  \item{mem_vector}{ vector of cluster membership }
  \item{adj_matrix}{ adjacency matrix }
}
\value{
  \item{permute_matrix}{a permuted matrix }}

\author{Mike Nowak \email{michael.nowak@gmail.com}}


\examples{
##


}
\keyword{ ~sna }

